streamlit-app/requirements.txt
streamlit
google-cloud-storage
google-cloud-aiplatform
google-cloud-bigquery
requests
langchain 
faiss-cpu 
sentence-transformers
pymupdf
sentence-transformers
faiss-cpu
langchain-text-splitters